tests/requirements.txt
oem-client-provider-package>=1.0.0
oem-format-json>=1.0.0
oem-storage-file>=1.0.0

pytest==2.9.1
pytest-cov==2.2.1